requirements.txt
mysql-connector-python==9.3.0
python-dotenv==1.1.0
SQLAlchemy==2.0.41
scikit-learn==1.6.1
pandas==2.2.3
fasttext==0.9.2
tensorflow==2.19.0